Add friends functionality

+++ app/controllers/CirclesController.php
@@ -152,36 +152,39 @@ class CirclesController extends \BaseController {
   public function update($id) {
     $input = Input::all();
     $circle = Circle::find($id);
+    if(Request::ajax()) {
+      // Handle add request
+      if(isset($input['add'])) {
+        $user = User::find($input['add']);
+        $circle->users()->attach($user);
 
-    // Handle add request
-    if(isset($input['add'])) {
-      $user = User::find($input['add']);
-      $circle->users()->attach($user);
+        $selector = "#circle-$id";
 
-      $selector = "#circle-$id";
+        $commands[] = array(
+          'method' => 'html',
+          'selector' => $selector,
+          'html' => utf8_encode(View::make('circles.item')->with(array(
+            'd' => $circle,
+          ))),
+        );
+      }
 
-      $commands[] = array(
-        'method' => 'html',
-        'selector' => $selector,
-        'html' => utf8_encode(View::make('circles.item')->with(array(
-          'd' => $circle,
-        ))),
-      );
+      // Handle delete request
+      if(isset($input['del'])) {
+        $user = User::find($input['del']);
+        $circle->users()->detach($user);
+
+        $commands[] = array(
+          'method' => 'remove',
+          'selector' => "#circle-$id #user-" . $user->id,
+        );
+      }
+
+      return Response::json($commands);
+
+    } else {
+      // TODO: some non AJAX actions
     }
-
-    // Handle delete request
-    if(isset($input['del'])) {
-      $user = User::find($input['del']);
-      $circle->users()->detach($user);
-
-      $commands[] = array(
-        'method' => 'remove',
-        'selector' => "#circle-$id #user-" . $user->id,
-      );
-
-    }
-
-    return Response::json($commands);
   }
 
   /**
